change in file with routing concept
from main.py file we are creating  a new folder routers and creating 2 new file as blog.py and user.py inside it.
