+++ Fastapi_app2/blog/schemas.py
@@ -1,10 +1,11 @@
 from typing import List 
 from pydantic import BaseModel
 
-class Blog(BaseModel):
-    body:str
+class BlogBase(BaseModel):
     title:str
-    
+    body:str
+
+class Blog(BlogBase):
     class Config:
         orm_mode=True
        
@@ -21,7 +22,7 @@ class User(BaseModel):
 class ShowUser(BaseModel):
     name:str
     email:str
-    blogs:List[Blog]=True
+    blogs:List[Blog]=[]
     class Config:
         orm_mode=True
